# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Portfolio Intelligence API for Indian equity investors
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /research/companies/{ticker}:
    get:
      operationId: getCompanyDetail
      tags: [research]
      summary: Full company research detail
      parameters:
        - name: ticker
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Company detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanyDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /research/decision-journal:
    get:
      operationId: getDecisionJournal
      tags: [research]
      summary: Investment decision journal entries
      responses:
        "200":
          description: Journal entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JournalEntry"
    post:
      operationId: addJournalEntry
      tags: [research]
      summary: Add journal entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JournalEntryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JournalEntry"

  # ── Market Intelligence ─────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]

    # ── Dashboard ──────────────────────────────────────────────────────────────
...
    Catalyst:
      type: object
      required: [id, ticker, name, event, date, impact, probability]
      properties:
        id:
          type: string
        ticker:
          type: string
        name:
          type: string
        event:
          type: string
        date:
          type: string
        impact:
          type: string
          enum: [high, medium, low]
        probability:
          type: string
          enum: [high, medium, low]
        notes:
          type: ["string", "null"]

    # ── Holdings ───────────────────────────────────────────────────────────────
...
    CompanyDetail:
      type: object
      required: [ticker, name, sector, marketCap, ltp, thesis, valuation, financialTrends, technicals, managementCommentary, catalysts, invalidationTriggers, earningsQualityFlags, lastUpdated]
      properties:
        ticker:
          type: string
        name:
          type: string
        sector:
          type: string
        marketCap:
          type: number
        ltp:
          type: number
        thesis:
          $ref: "#/components/schemas/Thesis"
        valuation:
          $ref: "#/components/schemas/ValuationCases"
        financialTrends:
          type: array
          items:
            $ref: "#/components/schemas/FinancialTrend"
        technicals:
          $ref: "#/components/schemas/Technicals"
        managementCommentary:
          type: array
          items:
            $ref: "#/components/schemas/ManagementComment"
        catalysts:
          type: array
          items:
            $ref: "#/components/schemas/Catalyst"
        invalidationTriggers:
          type: array
          items:
            $ref: "#/components/schemas/InvalidationTrigger"
        earningsQualityFlags:
          type: array
          items:
            $ref: "#/components/schemas/EarningsQualityFlag"
        lastUpdated:
          type: string
    Thesis:
      type: object
      required: [summary, bullCase, baseCase, bearCase, conviction, keyAssumptions, moatRating, managementRating, updatedAt]
      properties:
        summary:
          type: string
        bullCase:
          type: string
        baseCase:
          type: string
        bearCase:
          type: string
        conviction:
          type: string
          enum: [high, medium, low]
        keyAssumptions:
          type: array
          items:
            type: string
        moatRating:
          type: string
          enum: [wide, narrow, none]
        managementRating:
          type: string
          enum: [excellent, good, average, poor]
        updatedAt:
          type: string
    ValuationCases:
      type: object
      required: [bull, base, bear, currentPrice, upside, methodology]
      properties:
        bull:
          type: number
        base:
          type: number
        bear:
          type: number
        currentPrice:
          type: number
        upside:
          type: number
        methodology:
          type: string
    FinancialTrend:
      type: object
      required: [metric, periods, values, trend]
      properties:
        metric:
          type: string
        periods:
          type: array
          items:
            type: string
        values:
          type: array
          items:
            type: number
        trend:
          type: string
          enum: [improving, stable, deteriorating]
        unit:
          type: string
    Technicals:
      type: object
      required: [rsi14, ema50, ema200, support, resistance, macdSignal, volumeVsAvg20d, trend]
      properties:
        rsi14:
          type: number
        ema50:
          type: number
        ema200:
          type: number
        support:
          type: number
        resistance:
          type: number
        macdSignal:
          type: string
          enum: [bullish, bearish, neutral]
        volumeVsAvg20d:
          type: number
        trend:
          type: string
          enum: [uptrend, downtrend, sideways]
    ManagementComment:
      type: object
      required: [id, quarter, date, speaker, comment, sentiment, tags]
      properties:
        id:
          type: string
        quarter:
          type: string
        date:
          type: string
        speaker:
          type: string
        comment:
          type: string
        sentiment:
          type: string
          enum: [positive, neutral, negative, concerning]
        tags:
          type: array
          items:
            type: string
    InvalidationTrigger:
      type: object
      required: [id, trigger, severity, description]
      properties:
        id:
          type: string
        trigger:
          type: string
        severity:
          type: string
          enum: [critical, high, medium]
        description:
          type: string
        triggered:
          type: boolean
    EarningsQualityFlag:
      type: object
      required: [id, flag, severity, description, detected]
      properties:
        id:
          type: string
        flag:
          type: string
        severity:
          type: string
          enum: [red, amber, green]
        description:
          type: string
        detected:
          type: boolean
    JournalEntry:
      type: object
      required: [id, ticker, name, action, date, price, rationale, emotionalState, keyFactors, revisitDate]
      properties:
        id:
          type: string
        ticker:
          type: string
        name:
          type: string
        action:
          type: string
          enum: [buy, sell, hold, review]
        date:
          type: string
        price:
          type: number
        rationale:
          type: string
        emotionalState:
          type: string
          enum: [confident, anxious, uncertain, greedy, fearful, neutral]
        keyFactors:
          type: array
          items:
            type: string
        revisitDate:
          type: string
        outcome:
          type: ["string", "null"]
    JournalEntryInput:
      type: object
      required: [ticker, name, action, date, price, rationale, emotionalState, keyFactors, revisitDate]
      properties:
        ticker:
          type: string
        name:
          type: string
        action:
          type: string
          enum: [buy, sell, hold, review]
        date:
          type: string
        price:
          type: number
        rationale:
          type: string
        emotionalState:
          type: string
          enum: [confident, anxious, uncertain, greedy, fearful, neutral]
        keyFactors:
          type: array
          items:
            type: string
        revisitDate:
          type: string

    # ── Market ─────────────────────────────────────────────────────────────────
...