# lib/api-zod/src/generated/types/catalyst.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Portfolio Intelligence API for Indian equity investors
 * OpenAPI spec version: 0.1.0
 */
import type { CatalystImpact } from './catalystImpact';
import type { CatalystProbability } from './catalystProbability';

export interface Catalyst {
  id: string;
  ticker: string;
  name: string;
  event: string;
  date: string;
  impact: CatalystImpact;
  probability: CatalystProbability;
  /** @nullable */
  notes?: string | null;
}